ref(server): Move TUS helpers into separate module
Split `upload.rs` into `upload/mod.rs` (endpoint handler) and
`upload/tus.rs` (protocol constants, error types, metadata parsing).
